ENH: model utils restructure
Restructured the model_utils routines, grouping routines by type in a
separate submodule.
