+++ pysatModelUtils/utils/__init__.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (C) 2019, AGB & pysat team
+# Full license can be found in License.md
+#-----------------------------------------------------------------------------
+"""
+pysatModelUtils.utils
+=====================
+
+Utilities designed to extract, match, and compare modelled and observed data
+
+"""
+
+from __future__ import absolute_import,  unicode_literals
+
+from . import (extract)
+from . import (match)
+from . import (compare)
